feiaiagent_1.0-712c2a1c52a1aab73cfca7a545a36d75cf9a6920/requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
flask
flask-cors
websocket-client
Pillow>=9.0.0

